refactor: useAuthRedirect hook and change directory name
