+++ client/src/components/Chat/Input/GenerationButtons.tsx
