+++ client/src/routes/useAuthRedirect.ts
@@ -1,0 +1,24 @@
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useAuthContext } from '~/hooks';
+
+export default function useAuthRedirect() {
+  const { isAuthenticated } = useAuthContext();
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      if (!isAuthenticated) {
+        navigate('/login', { replace: true });
+      }
+    }, 300);
+
+    return () => {
+      clearTimeout(timeout);
+    };
+  }, [isAuthenticated, navigate]);
+
+  return {
+    isAuthenticated,
+  };
+}
